pyproject.toml
[tool.poetry]
name = "itmx-kaggle"
version = "0.1.0"
description = "Data pipeline with DBT, DuckDB, Prefect and PySpark"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
dbt-duckdb = "^1.7.0"
prefect = "^2.14.0"
duckdb = "^0.10.0"
pyspark = "^3.5.0"
pandas = "^2.2.0"
pyarrow = "^15.0.0"
sqlalchemy = "^2.0.0"
python-dotenv = "^1.0.0"
networkx = "^3.1"
scipy = "^1.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
ruff = "^0.3.0"
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"